Make the Log\Mailer and Http\ContentGuard optionally and move them on App Extensions
