+++ app/Events.php
@@ -14,7 +14,7 @@ use Forensics\Console;
 /** Define Events. */
 
 // Add a Listener Class to the Event 'test'.
-Event::listen('test', 'App\Events\Test@handle');
+Event::listen('test', 'App\Extensions\Events\Test@handle');
 
 // Add a Listener Closure to the Event 'test'.
 Event::listen('test', function($data) {
