+++ app/Extensions/Log/Mailer.php
@@ -5,7 +5,7 @@
  * @author Virgil-Adrian Teaca - virgil@giulianaeassociati.com
  * @version 3.0
  */
-namespace Log;
+namespace App\Extensions\Log;
 
 use Foundation\Application;
